Revert "Mobile blog"
