+++ components/blog/blogDesktop/articlePreview/excerpt/Excerpt.js
@@ -2,9 +2,8 @@ import React from "react";
 import styled from "styled-components";
 
 const ExcerptContainer = styled.p`
-  color: #a8aaac;
-  font-size: clamp(0.9rem, 1vw, 1rem);
-  max-width: 55rem;
+  color: ${(p) => p.theme.colors.white300};
+  font-size: 1em;
 `;
 
 const Excerpt = ({ content }) => {
