+++ package.json
@@ -27,7 +27,6 @@
     "react-icons": "^4.3.1",
     "react-mailchimp-subscribe": "^2.1.3",
     "react-tweet-embed": "^2.0.0",
-    "react-web-share": "^1.1.6",
     "styled-components": "^5.3.3"
   },
   "devDependencies": {
